Repackage classes and tests into separate chessboard and piece packages
