+++ ChessProject-Java/src/main/java/com/solarwindsmsp/chess/piece/Pawn.java
@@ -1,4 +1,6 @@
-package com.solarwindsmsp.chess;
+package com.solarwindsmsp.chess.piece;
+
+import com.solarwindsmsp.chess.chessboard.ChessBoard;
 
 public class Pawn {
