+++ ChessProject-Java/src/main/java/com/solarwindsmsp/chess/piece/PieceColor.java
@@ -1,4 +1,4 @@
-package com.solarwindsmsp.chess;
+package com.solarwindsmsp.chess.piece;
 
 public enum PieceColor {
