+++ ChessProject-Java/src/test/java/com/solarwindsmsp/chess/piece/PawnTest.java
@@ -1,5 +1,6 @@
-package com.solarwindsmsp.chess;
+package com.solarwindsmsp.chess.piece;
 
+import com.solarwindsmsp.chess.chessboard.ChessBoard;
 import org.junit.Before;
 import org.junit.Test;
